Unais

+++ Server/Client/User/UserMaster.js
@@ -36,7 +36,7 @@ export const getUsers = async (req, res) => {
         message: "successfull",
         status: 200,
       });
-      // console.log(result);
+      //console.log(result);
     } else {
       res.status(200).json({
         data: [
@@ -59,7 +59,7 @@ export const getUsers = async (req, res) => {
         message: "no data found",
       });
     }
-    console.log(result);
+    //console.log(result);
   } catch (error) {
     console.log(error);
   }
@@ -90,7 +90,7 @@ export const getRights = async (req, res) => {
         message: "successfull",
         status: 200,
       });
-      // console.log(result);
+      //console.log(result);
     } else {
       res.status(200).json({
         data: [
@@ -137,11 +137,7 @@ export const CreateUser = async (req, res) => {
         newData.UserLocation,
         Adddate,
         newData.AddedBy,
-        newData.UserType == 3
-          ? rights
-          : newData.UserType == 6
-          ? rights
-          : newData.RightsDetails,
+        newData.UserType == 3 ? rights : newData.UserType == 6 ? rights : newData.RightsDetails,
       ]
     );
     if (result.length > 0) {
@@ -188,7 +184,7 @@ export const CreateUser = async (req, res) => {
         message: "no data found",
         status: 404,
       });
-      console.log(result);
+      //console.log(result);
     }
   } catch (error) {
     console.log(error);
@@ -203,10 +199,7 @@ export const getSpecificUser = async (req, res) => {
   }
   const id = req.params.id;
   try {
-    const result = await pool.query(
-      "SELECT * FROM adminuserlogins WHERE LoginID = ?; ",
-      [id]
-    );
+    const result = await pool.query("SELECT * FROM adminuserlogins WHERE LoginID = ?; ", [id]);
     if (result[0].length > 0) {
       const data = encodePasswords(result[0]);
 
@@ -231,7 +224,7 @@ export const getSpecificUser = async (req, res) => {
         message: "successfull",
         status: 200,
       });
-      // console.log(result);
+      //console.log(result);
     } else {
       res.json({
         data: [
@@ -255,7 +248,7 @@ export const getSpecificUser = async (req, res) => {
         status: 200,
       });
     }
-    // console.log(result);
+    //console.log(result);
   } catch (error) {
     console.log(error);
   }
@@ -279,11 +272,7 @@ export const updateUser = async (req, res) => {
         newData.UserLocation,
         ModifiedDate,
         newData.ModifiedBy,
-        newData.UserType == 3
-          ? rights
-          : newData.UserType == 6
-          ? rights
-          : newData.RightsDetails,
+        newData.UserType == 3 ? rights : newData.UserType == 6 ? rights : newData.RightsDetails,
         id,
       ]
     );
@@ -341,40 +330,28 @@ export const updateUserSubjects = async (req, res) => {
   const userClass = req.body.class;
   const userSubjects = req.body.subjects;
   console.log(userClass, id, userSubjects);
-  const values = userSubjects.map((subjectId, index) => [
-    subjectId,
-    userClass[index],
-    id,
-  ]);
+  const values = userSubjects.map((subjectId, index) => [subjectId, userClass[index], id]);
   const insertRowQuery = `
   INSERT INTO adminusersubjects (SubjectID, ClassId, LoginID) VALUES ?
 `;
   if (userClass.length > 0 && userSubjects.length > 0) {
     try {
-      const deleteAll = await pool.query(
-        "DELETE FROM adminusersubjects WHERE LoginiD = ?",
-        [id],
-        (err, results) => {
-          if (err) {
-            return connection.rollback(() => {
-              console.error("Error deleting rows:", err);
-              throw err;
-            });
-          }
+      const deleteAll = await pool.query("DELETE FROM adminusersubjects WHERE LoginiD = ?", [id], (err, results) => {
+        if (err) {
+          return connection.rollback(() => {
+            console.error("Error deleting rows:", err);
+            throw err;
+          });
         }
-      );
-      const result = await pool.query(
-        insertRowQuery,
-        [values],
-        (err, results, fields) => {
-          if (err) {
-            console.error("Error inserting rows:", err);
-            return;
-          }
-          console.log("Rows inserted successfully");
+      });
+      const result = await pool.query(insertRowQuery, [values], (err, results, fields) => {
+        if (err) {
+          console.error("Error inserting rows:", err);
+          return;
         }
-      );
-      console.log(result);
+        console.log("Rows inserted successfully");
+      });
+      //console.log(result);
       if (result.length > 0) {
         res.json({
           data: [
@@ -431,27 +408,19 @@ export const createUserSubjects = async (req, res) => {
   const id = req.body.id;
   const userClass = req.body.class;
   const userSubjects = req.body.subjects;
-  const values = userSubjects.map((subjectId, index) => [
-    subjectId,
-    userClass[index],
-    id,
-  ]);
+  const values = userSubjects.map((subjectId, index) => [subjectId, userClass[index], id]);
   const insertRowQuery = `
     INSERT INTO adminusersubjects (SubjectID, ClassId, LoginID) VALUES ?
   `;
   if (userClass.length > 0 && userSubjects.length > 0) {
     try {
-      const result = await pool.query(
-        insertRowQuery,
-        [values],
-        (err, results, fields) => {
-          if (err) {
-            console.error("Error inserting rows:", err);
-            return;
-          }
-          console.log("Rows inserted successfully");
+      const result = await pool.query(insertRowQuery, [values], (err, results, fields) => {
+        if (err) {
+          console.error("Error inserting rows:", err);
+          return;
         }
-      );
+        console.log("Rows inserted successfully");
+      });
       if (result.length > 0) {
         res.json({
           data: [
@@ -508,10 +477,7 @@ export const getUsersSubject = async (req, res) => {
   const id = req.params.id;
   console.log(id);
   try {
-    const result = await pool.query(
-      "SELECT * FROM adminusersubjects WHERE LoginId = ?",
-      [id]
-    );
+    const result = await pool.query("SELECT * FROM adminusersubjects WHERE LoginId = ?", [id]);
 
     if (result[0].length > 0) {
       res.json({
@@ -535,7 +501,7 @@ export const getUsersSubject = async (req, res) => {
         message: "successfull",
         status: 200,
       });
-      // console.log(result);
+      //console.log(result);
     } else {
       res.status(200).json({
         data: [
@@ -558,7 +524,7 @@ export const getUsersSubject = async (req, res) => {
         message: "no data found",
       });
     }
-    console.log(result);
+    //console.log(result);
   } catch (error) {
     console.log(error);
   }
@@ -571,177 +537,199 @@ export const importStudentData = async (req, res) => {
   formData.append("title", "GetStudentMastersForQB");
   formData.append("description", "QBStudentMasterList");
 
-  const result = await axios.post(url, formData);
-
-  const intialData = JSON.parse(
-    result.data.substring(0, result.data.indexOf("||JasonEnd", 0))
-  );
-  let studentMaster = JSON.parse(
-    intialData.map((item) =>
-      item.JSONData1.substring(0, result.data.indexOf("||JasonEnd", 0))
-    )
-  );
-  let studentYear = JSON.parse(
-    intialData.map((item) =>
-      item.JSONData2.substring(0, result.data.indexOf("||JasonEnd", 0))
-    )
-  );
-
   try {
-    if (studentMaster.length > 0) {
-      const insertRowQuery = `
-    INSERT INTO studentmaster (StudentID,STDMSTID, ADMNO, SNAME,FATHERNAME,GENDER,DOB,DOA,SMSPHONE,ADDRESS,MNAME,TCSTS,PASSWORD,VERCODE) VALUES ?
-  `;
-      const dataArray = studentMaster.map((student) => [
-        student.StudentID,
-        student.STDMSTID,
-        student.ADMNO,
-        student.SNAME,
-        student.FATHERNAME,
-        student.GENDER,
-        new Date(student.DOB),
-        new Date(student.DOA),
-        student.SMSPHONE,
-        student.ADDRESS,
-        student.MNAME,
-        student.TCSTS,
-        student.PASSWORD,
-        student.VERCODE,
-      ]);
+    const result = await axios.post(url, formData);
+
+    // console.log(result, "result----------------");
+
+    const intialData = JSON.parse(result.data.substring(0, result.data.indexOf("||JasonEnd", 0)));
+    let studentMaster = JSON.parse(intialData.map((item) => item.JSONData1.substring(0, result.data.indexOf("||JasonEnd", 0))));
+    let studentYear = JSON.parse(intialData.map((item) => item.JSONData2.substring(0, result.data.indexOf("||JasonEnd", 0))));
+
+    try {
       const deletetable = await pool.query("DELETE FROM studentmaster");
 
-      const result =await pool.query(
-        insertRowQuery,
-        [dataArray],
-        (error, results, fields) => {
-          if (error) {
-            console.error("Error inserting data:", error);
-            return;
-          }
-          console.log("Data inserted successfully:", results);
+      const resultList = [];
+      if (studentMaster.length > 0) {
+        const insertRowQuery = `
+    INSERT INTO studentmaster (StudentID,STDMSTID, ADMNO, SNAME,FATHERNAME,GENDER,DOB,DOA,SMSPHONE,ADDRESS,MNAME,TCSTS,PASSWORD,VERCODE) VALUES ?
+  `;
+        // const dataArray = studentMaster.map(async (student) => {
+        //   const stu = [
+        //     student.StudentID,
+        //     student.STDMSTID,
+        //     student.ADMNO,
+        //     student.SNAME,
+        //     student.FATHERNAME,
+        //     student.GENDER,
+        //     new Date(student.DOB),
+        //     new Date(student.DOA),
+        //     student.SMSPHONE,
+        //     student.ADDRESS,
+        //     student.MNAME,
+        //     student.TCSTS,
+        //     student.PASSWORD,
+        //     student.VERCODE,
+        //   ];
+        // //   console.log(stu, "00000000000000000000000000000000000000000000000");
+        //   const result = await pool.query(insertRowQuery, stu, (error, results, fields) => {
+        //     if (error) {
+        //       console.error("Error inserting data:", error);
+        //       return;
+        //     }
+        //     console.log("Data inserted successfully:", results,'---------------------');
+        //   });
+        //   resultList.push(result)
+        // });
+        // console.log(resultList, "###############################################");
+        // console.log(studentMaster.length, "length");
+        const dataArray = studentMaster.map((student) => [
+          student.StudentID,
+          student.STDMSTID,
+          student.ADMNO,
+          student.SNAME,
+          student.FATHERNAME,
+          student.GENDER,
+          new Date(student.DOB),
+          new Date(student.DOA),
+          student.SMSPHONE,
+          student.ADDRESS,
+          student.MNAME,
+          student.TCSTS,
+          student.PASSWORD,
+          student.VERCODE,
+        ]);
+
+        // Perform the batch insertion
+        try {
+          const [result] = await pool.query(insertRowQuery, [dataArray]);
+          //   console.log("Data inserted successfully:", result);
+          resultList.push(result);
+        } catch (error) {
+          console.error("Error inserting data:", error);
         }
-      );
-      if (result.length > 0) {
-        res.json({
-          data: [
-            {
-              ActionType: "",
-              ErrorMessage: "",
-              ErrorCode: "",
-              JSONData1: result,
-              JSONData2: [],
-              JSONData3: [],
-              JSONData4: [],
-              JSONData5: [],
-              JSONData1Remarks: "",
-              JSONData2Remarks: "",
-              JSONData3Remarks: "",
-              JSONData4Remarks: "",
-              JSONData5Remarks: "",
-            },
-          ],
-          message: "successfull",
-          status: 200,
-        });
-      } else {
-        res.json({
-          data: [
-            {
-              ActionType: "",
-              ErrorMessage: "",
-              ErrorCode: "",
-              JSONData1: [result],
-              JSONData2: [],
-              JSONData3: [],
-              JSONData4: [],
-              JSONData5: [],
-              JSONData1Remarks: "",
-              JSONData2Remarks: "",
-              JSONData3Remarks: "",
-              JSONData4Remarks: "",
-              JSONData5Remarks: "",
-            },
-          ],
-          message: "no data found",
-          status: 404,
-        });
+        // console.log(resultList, "list %%%%%%%%%%%%%%%%%%%%%%%%%%%");
+
+        if (resultList.length > 0) {
+          res.json({
+            data: [
+              {
+                ActionType: "",
+                ErrorMessage: "",
+                ErrorCode: "",
+                JSONData1: resultList,
+                JSONData2: [],
+                JSONData3: [],
+                JSONData4: [],
+                JSONData5: [],
+                JSONData1Remarks: "",
+                JSONData2Remarks: "",
+                JSONData3Remarks: "",
+                JSONData4Remarks: "",
+                JSONData5Remarks: "",
+              },
+            ],
+            message: "successfull",
+            status: 200,
+          });
+        } else {
+          res.json({
+            data: [
+              {
+                ActionType: "",
+                ErrorMessage: "",
+                ErrorCode: "",
+                JSONData1: [resultList],
+                JSONData2: [],
+                JSONData3: [],
+                JSONData4: [],
+                JSONData5: [],
+                JSONData1Remarks: "",
+                JSONData2Remarks: "",
+                JSONData3Remarks: "",
+                JSONData4Remarks: "",
+                JSONData5Remarks: "",
+              },
+            ],
+            message: "no data found",
+            status: 404,
+          });
+        }
       }
-    }
-    if (studentYear.length > 0) {
-      const insertRowQuery = `
+      if (studentYear.length > 0) {
+        const insertRowQuery = `
     INSERT INTO studentyearmaster (StudentYearID,StudentID, STDYRID, STDMSTID,CLID,RNO,CLNAME,SECNAME,ACASTART,CLASSCAT,STS) VALUES ?
   `;
-      const dataArray = studentYear.map((student) => [
-        student.StudentYearID,
-        student.StudentID,
-        student.STDYRID,
-        student.STDMSTID,
-        student.CLID,
-        student.RNO,
-        student.CLNAME,
-        student.SECNAME,
-        student.ACASTART,
-        student.CLASSCAT,
-        student.STS,
-      ]);
+        const dataArray = studentYear.map((student) => [
+          student.StudentYearID,
+          student.StudentID,
+          student.STDYRID,
+          student.STDMSTID,
+          student.CLID,
+          student.RNO,
+          student.CLNAME,
+          student.SECNAME,
+          student.ACASTART,
+          student.CLASSCAT,
+          student.STS,
+        ]);
 
-      const clearTable = await pool.query("DELETE FROM studentyearmaster");
-      const result =await pool.query(
-        insertRowQuery,
-        [dataArray],
-        (error, results, fields) => {
+        const clearTable = await pool.query("DELETE FROM studentyearmaster");
+        const result = await pool.query(insertRowQuery, [dataArray], (error, results, fields) => {
           if (error) {
             console.error("Error inserting data:", error);
             return;
           }
           console.log("Data inserted successfully:", results);
+        });
+        if (result.length > 0) {
+          res.json({
+            data: [
+              {
+                ActionType: "",
+                ErrorMessage: "",
+                ErrorCode: "",
+                JSONData1: result,
+                JSONData2: [],
+                JSONData3: [],
+                JSONData4: [],
+                JSONData5: [],
+                JSONData1Remarks: "",
+                JSONData2Remarks: "",
+                JSONData3Remarks: "",
+                JSONData4Remarks: "",
+                JSONData5Remarks: "",
+              },
+            ],
+            message: "successfull",
+            status: 200,
+          });
+        } else {
+          res.json({
+            data: [
+              {
+                ActionType: "",
+                ErrorMessage: "",
+                ErrorCode: "",
+                JSONData1: [result],
+                JSONData2: [],
+                JSONData3: [],
+                JSONData4: [],
+                JSONData5: [],
+                JSONData1Remarks: "",
+                JSONData2Remarks: "",
+                JSONData3Remarks: "",
+                JSONData4Remarks: "",
+                JSONData5Remarks: "",
+              },
+            ],
+            message: "no data found",
+            status: 404,
+          });
         }
-      );
-      if (result.length > 0) {
-        res.json({
-          data: [
-            {
-              ActionType: "",
-              ErrorMessage: "",
-              ErrorCode: "",
-              JSONData1: result,
-              JSONData2: [],
-              JSONData3: [],
-              JSONData4: [],
-              JSONData5: [],
-              JSONData1Remarks: "",
-              JSONData2Remarks: "",
-              JSONData3Remarks: "",
-              JSONData4Remarks: "",
-              JSONData5Remarks: "",
-            },
-          ],
-          message: "successfull",
-          status: 200,
-        });
-      } else {
-        res.json({
-          data: [
-            {
-              ActionType: "",
-              ErrorMessage: "",
-              ErrorCode: "",
-              JSONData1: [result],
-              JSONData2: [],
-              JSONData3: [],
-              JSONData4: [],
-              JSONData5: [],
-              JSONData1Remarks: "",
-              JSONData2Remarks: "",
-              JSONData3Remarks: "",
-              JSONData4Remarks: "",
-              JSONData5Remarks: "",
-            },
-          ],
-          message: "no data found",
-          status: 404,
-        });
       }
+    } catch (error) {
+      console.log(error);
     }
   } catch (error) {
     console.log(error);
